Split up the btcd connector from the blockchain explorer.
Makes the code easier to figure out.
